minor, use identity slot in messages to prevent router/dealer loops
The idea is to pass a random 32b number around.  The probability that bit I
is set shall be 50% ideally.  Every router |s its own random cookie onto
the cookie read.  So it's 16 bits free after the first hop and 8 bits free
after the second hop, and so forth.  So roughly we can detect 4-hop loops
safely.

Signed-off-by: Sebastian Freundt <freundt@ga-group.nl>